refactored cli args
