+++ src/auth.ts
@@ -8,11 +8,11 @@ import { AxiosError } from 'axios';
 import login from 'metacall-protocol/login';
 import API from 'metacall-protocol/protocol';
 import { expiresIn } from 'metacall-protocol/token';
+import args from './cli/args';
 import { input, maskedInput } from './cli/inputs';
 import { info, warn } from './cli/messages';
 import { loginSelection } from './cli/selection';
 import { Config, save } from './config';
-import { args } from './index';
 import { forever, opt } from './utils';
 
 const authToken = async (config: Config): Promise<string> => {
